Updated the parser and code generator for tests
This is a checkpoint.
Testing for the code generator required changeds in both parser.dart and
code_generator.dart to be made. Currently, an empty class works. An
empty method *did* work, however, upon changing things for the third
test (method with statements), things changed and it currently doesnt
work. To be continued...